move test case
move onadata.apps.api.tests.viewsets.test_xform_submission_viewset.TestXFormSubmissionViewSet.test_registration_form_inactive to onadata.apps.logger.tests.models.test_instance.TestInstance.test_registration_form_inactive
